Former-commit-id: 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f [formerly 56dc338129ed46612df4db1926e0b7e91551b60d]]
Former-commit-id: adfeb1306751695abe2eaaf23d62ed4f3baa2430
Former-commit-id: 12c854e7cc66f4fd55c067ee414a3797e0cac5d2

+++ edexOsgi/com.raytheon.edex.plugin.obs/src/com/raytheon/edex/plugin/obs/metar/MetarDecoder.java
@@ -83,6 +83,8 @@ import com.vividsolutions.jts.geom.impl.CoordinateArraySequence;
  * Sep 17, 2013 2378        njensen     Improve 3/6 hr precip decoding
  * May 12, 2014 DR 17151    D. Friedman Fix 6hr min/max temp decoding.
  * May 14, 2014 2536        bclement    moved WMO Header to common, removed TimeTools usage
+ * Jul 23, 2014  3410       bclement    location changed to floats
+ * 
  * </pre>
  * 
  * @author bphillip
@@ -288,8 +290,8 @@ public class MetarDecoder extends AbstractDecoder {
                     ObStation station = getStationInfo(icao);
                     if (station != null) {
                         SurfaceObsLocation loc = new SurfaceObsLocation(icao);
-                        Double lat = station.getGeometry().getY();
-                        Double lon = station.getGeometry().getX();
+                        float lat = (float) station.getGeometry().getY();
+                        float lon = (float) station.getGeometry().getX();
                         loc.assignLocation(lat, lon);
                         loc.setElevation(station.getElevation());
