+++ edexOsgi/com.raytheon.uf.common.dataplugin.acarssounding/src/com/raytheon/uf/common/dataplugin/acarssounding/ACARSSoundingLayer.java
@@ -1,19 +1,19 @@
 /**
  * This software was developed and / or modified by Raytheon Company,
  * pursuant to Contract DG133W-05-CQ-1067 with the US Government.
- * 
+ *
  * U.S. EXPORT CONTROLLED TECHNICAL DATA
  * This software product contains export-restricted data whose
  * export/transfer/disclosure is restricted by U.S. law. Dissemination
  * to non-U.S. persons whether in the United States or abroad requires
  * an export license or other authorization.
- * 
+ *
  * Contractor Name:        Raytheon Company
  * Contractor Address:     6825 Pine Street, Suite 340
  *                         Mail Stop B8
  *                         Omaha, NE 68106
  *                         402.291.0100
- * 
+ *
  * See the AWIPS II Master Rights File ("Master Rights File.pdf") for
  * further licensing information.
  **/
@@ -41,17 +41,18 @@ import com.vividsolutions.jts.geom.Point;
 
 /**
  * Layer for an ACARS Sounding
- * 
+ *
  * <pre>
- * 
+ *
  * SOFTWARE HISTORY
  * Date         Ticket#    Engineer    Description
  * ------------ ---------- ----------- --------------------------
  * Apr 30, 2009            jkorman     Initial creation
  * Oct 22, 2013 2361       njensen     Remove XML annotations
- * 
+ * Jul 22, 2014 3392       nabowle     Change Double fields to Float.
+ *
  * </pre>
- * 
+ *
  * @author jkorman
  * @version 1.0
  */
@@ -93,23 +94,23 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     // Observation air temperature in degrees Kelvin.
     @Column
     @DynamicSerializeElement
-    private Double temp;
+    private Float temp;
 
     // Observation dewpoint temperature in degrees Kelvin.
     // Decimal(5,2)
     @Column
     @DynamicSerializeElement
-    private Double dwpt;
+    private Float dwpt;
 
     // Relative Humidity in percent. Decimal(5,2)
     @Column
     @DynamicSerializeElement
-    private Double humidity;
+    private Float humidity;
 
     // Mixing ratio in Kg/Kg
     @Column
     @DynamicSerializeElement
-    private Double mixingRatio;
+    private Float mixingRatio;
 
     // Observation wind direction in angular degrees. Integer
     @Column
@@ -120,7 +121,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     // Decimal(5,2)
     @Column
     @DynamicSerializeElement
-    private Double windSpeed;
+    private Float windSpeed;
 
     // Base height of reported icing.
     @Column
@@ -155,10 +156,10 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     // Indicated or calculated pressure.
     @Column
     @DynamicSerializeElement
-    private Double pressure;
+    private Float pressure;
 
     /**
-     * 
+     *
      */
     public ACARSSoundingLayer() {
 
@@ -166,7 +167,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * The record id for this entry. Set from Hibernate.
-     * 
+     *
      * @return the recordId
      */
     public Integer getRecordId() {
@@ -175,7 +176,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Set the record id. Should be set by Hibernate only!
-     * 
+     *
      * @param recordId
      *            the recordId to set
      */
@@ -185,7 +186,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Get the parent instance that contains this level information.
-     * 
+     *
      * @return The parent instance.
      */
     public ACARSSoundingRecord getParent() {
@@ -194,7 +195,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Set the parent instance that contains this level information.
-     * 
+     *
      * @param parent
      *            The parent instance.
      */
@@ -233,7 +234,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     }
 
     /**
-     * 
+     *
      * @return
      */
     public AircraftObsLocation getLocation() {
@@ -241,7 +242,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     }
 
     /**
-     * 
+     *
      * @param location
      */
     public void setLocation(AircraftObsLocation location) {
@@ -250,7 +251,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Get this observation's geometry.
-     * 
+     *
      * @return The geometry for this observation.
      */
     public Geometry getGeometry() {
@@ -259,7 +260,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Get the geometry latitude.
-     * 
+     *
      * @return The geometry latitude.
      */
     public double getLatitude() {
@@ -268,7 +269,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Get the geometry longitude.
-     * 
+     *
      * @return The geometry longitude.
      */
     public double getLongitude() {
@@ -277,7 +278,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Get the elevation, in meters, of the observing platform or location.
-     * 
+     *
      * @return The observation elevation, in meters.
      */
     public Boolean getLocationDefined() {
@@ -286,7 +287,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Get the elevation, in meters, of the observing platform or location.
-     * 
+     *
      * @return The observation elevation, in meters.
      */
     public Integer getFlightLevel() {
@@ -294,7 +295,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     }
 
     /**
-     * 
+     *
      * @return
      */
     public String getFlightNumber() {
@@ -319,7 +320,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     /**
      * @return the temp
      */
-    public Double getTemp() {
+    public Float getTemp() {
         return temp;
     }
 
@@ -327,14 +328,14 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
      * @param temp
      *            the temp to set
      */
-    public void setTemp(Double temp) {
+    public void setTemp(Float temp) {
         this.temp = temp;
     }
 
     /**
      * @return the dwpt
      */
-    public Double getDwpt() {
+    public Float getDwpt() {
         return dwpt;
     }
 
@@ -342,14 +343,14 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
      * @param dwpt
      *            the dwpt to set
      */
-    public void setDwpt(Double dwpt) {
+    public void setDwpt(Float dwpt) {
         this.dwpt = dwpt;
     }
 
     /**
      * @return the humidity
      */
-    public Double getHumidity() {
+    public Float getHumidity() {
         return humidity;
     }
 
@@ -357,14 +358,14 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
      * @param humidity
      *            the humidity to set
      */
-    public void setHumidity(Double humidity) {
+    public void setHumidity(Float humidity) {
         this.humidity = humidity;
     }
 
     /**
      * @return the mixingRatio
      */
-    public Double getMixingRatio() {
+    public Float getMixingRatio() {
         return mixingRatio;
     }
 
@@ -372,7 +373,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
      * @param mixingRatio
      *            the mixingRatio to set
      */
-    public void setMixingRatio(Double mixingRatio) {
+    public void setMixingRatio(Float mixingRatio) {
         this.mixingRatio = mixingRatio;
     }
 
@@ -394,7 +395,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     /**
      * @return the windSpeed
      */
-    public Double getWindSpeed() {
+    public Float getWindSpeed() {
         return windSpeed;
     }
 
@@ -402,7 +403,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
      * @param windSpeed
      *            the windSpeed to set
      */
-    public void setWindSpeed(Double windSpeed) {
+    public void setWindSpeed(Float windSpeed) {
         this.windSpeed = windSpeed;
     }
 
@@ -499,7 +500,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
     /**
      * @return the pressure
      */
-    public Double getPressure() {
+    public Float getPressure() {
         return pressure;
     }
 
@@ -507,12 +508,12 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
      * @param pressure
      *            the pressure to set
      */
-    public void setPressure(Double pressure) {
+    public void setPressure(Float pressure) {
         this.pressure = pressure;
     }
 
     /**
-     * 
+     *
      */
     @Override
     public ISpatialObject getSpatialObject() {
@@ -521,7 +522,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /**
      * Provide a string representation of this observation.
-     * 
+     *
      * @return
      */
     @Override
@@ -532,7 +533,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /*
      * (non-Javadoc)
-     * 
+     *
      * @see java.lang.Object#hashCode()
      */
     @Override
@@ -549,7 +550,7 @@ public class ACARSSoundingLayer implements Serializable, ISpatialEnabled {
 
     /*
      * (non-Javadoc)
-     * 
+     *
      * @see java.lang.Object#equals(java.lang.Object)
      */
     @Override
